# api/openapi.yaml
openapi: 3.0.3
info:
  title: openbot API
  version: 0.1.0
  description: |
    HTTP API for openbotd. Authenticated with a session cookie (`openbot_session`).
    Mutating requests must send `X-Openbot-CSRF: 1` (checked together with the Origin header).

    Realtime updates are delivered over a WebSocket at `GET /api/ws`; each frame is a JSON
    `WsEvent` (see components). The WebSocket is not described as an operation because
    OpenAPI can't express it; its payloads are.
servers:
  - url: /api
paths:
...
  /onboarding:
    get:
      operationId: getOnboarding
      tags: [onboarding]
      description: What's left to set up after the account exists.
      responses:
        "200":
          description: Onboarding state
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OnboardingState" }
        "401": { $ref: "#/components/responses/Error" }
  /onboarding/complete:
    post:
      operationId: completeOnboarding
      tags: [onboarding]
      description: |
        Create the starter agent with the chosen model. Requires the provider to be configured.
        Only allowed once (while no agents exist).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CompleteOnboardingRequest" }
      responses:
        "200":
          description: Starter agent created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CompleteOnboardingResponse" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /settings/provider:
    get:
      operationId: getProviderSettings
      tags: [settings]
      responses:
        "200":
          description: Model provider settings (the key itself is never returned)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProviderSettings" }
        "401": { $ref: "#/components/responses/Error" }
    put:
      operationId: updateProviderSettings
      tags: [settings]
      description: Validate the API key against the provider, then save it (encrypted at rest).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateProviderSettingsRequest" }
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProviderSettings" }
        "400":
          description: The provider rejected the key, or the request is invalid
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "502": { $ref: "#/components/responses/Error" }
...
  /agents:
    get:
      operationId: listAgents
      tags: [agents]
      responses:
        "200":
          description: All agents
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Agent" }
        "401": { $ref: "#/components/responses/Error" }
...
  /connectors/types:
    get:
      operationId: listConnectorTypes
      tags: [connectors]
      description: The kinds of apps that can be connected, with the fields each needs.
      responses:
        "200":
          description: Connector types
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ConnectorType" }
        "401": { $ref: "#/components/responses/Error" }
  /connectors:
    get:
      operationId: listConnectors
      tags: [connectors]
      responses:
        "200":
          description: Connected accounts
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Connector" }
        "401": { $ref: "#/components/responses/Error" }
    post:
      operationId: createConnector
      tags: [connectors]
      description: Connect an account. Credentials are checked with the service before saving.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateConnectorRequest" }
      responses:
        "201":
          description: Connected
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Connector" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "502": { $ref: "#/components/responses/Error" }
...
  /connectors/{connectorId}:
    parameters:
      - name: connectorId
        in: path
        required: true
        schema: { type: string }
    patch:
      operationId: updateConnector
      tags: [connectors]
      description: |
        Rename, replace credentials (empty values keep the saved ones), change config, or set
        which agents may use the account.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateConnectorRequest" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Connector" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "502": { $ref: "#/components/responses/Error" }
    delete:
      operationId: deleteConnector
      tags: [connectors]
      responses:
        "204": { description: Disconnected }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
...
  /messages/{messageId}/connect:
    parameters:
      - name: messageId
        in: path
        required: true
        schema: { type: string }
    post:
      operationId: connectPrompt
      tags: [chats]
      description: |
        Accept a connection an agent proposed (a prompt of kind connect). openbot checks the
        credentials with the app, saves the connection with the proposed agents' access, marks
        the prompt answered (broadcast as message.updated) and tells the agent. Credentials go
        straight to the connection and are never shown to the agent. To decline, answer the
        prompt with option 1.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ConnectPromptRequest" }
      responses:
        "200":
          description: The updated message
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Message" }
        "400":
          description: The app rejected the details; the prompt stays pending
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409":
          description: The prompt was already answered or dismissed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "502":
          description: openbot couldn't reach the app
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: openbot_session
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [message]
      properties:
        message: { type: string }
        code:
          type: string
          description: |
            Machine-readable reason where clients react to it: sign_in_required (an MCP server
            wants a sign-in instead of a key), reconnect (a sign-in expired)
...
    OnboardingState:
      type: object
      required: [providerConfigured, starterAgentCreated, completed]
      properties:
        providerConfigured:
          type: boolean
          description: An OpenCode Go API key has been saved
        starterAgentCreated: { type: boolean }
        completed:
          type: boolean
          description: Both steps are done; the app is ready to use
    CompleteOnboardingRequest:
      type: object
      required: [model]
      properties:
        model:
          type: string
          description: Model id for the starter agent (from GET /models)
        agentName:
          type: string
          minLength: 1
          maxLength: 64
          description: Starter agent name; defaults to "openbot"
    CompleteOnboardingResponse:
      type: object
      required: [agent, chatId]
      properties:
        agent: { $ref: "#/components/schemas/Agent" }
        chatId:
          type: string
          description: The starter agent's DM
    ProviderSettings:
      type: object
      required: [provider, configured, keyHint]
      properties:
        provider: { type: string, enum: [opencode-go] }
        configured: { type: boolean }
        keyHint:
          type: string
          nullable: true
          description: Last 4 characters of the saved key, e.g. "…a1b2"
    UpdateProviderSettingsRequest:
      type: object
      required: [apiKey]
      properties:
        apiKey: { type: string, minLength: 1, maxLength: 512 }
    Agent:
      type: object
      required: [id, name, instructions, model, language, notifications, trustMode, isAdmin, activity, createdAt]
      properties:
        id: { type: string }
        name: { type: string }
        instructions:
          type: string
          description: What the agent is for and how it should work (its own system instructions)
        model: { type: string }
        language: { type: string, description: 'Reply language, or "auto"' }
        notifications: { type: boolean }
        trustMode: { type: string, enum: [ask, trusted] }
        isAdmin: { type: boolean }
        activity: { $ref: "#/components/schemas/AgentActivity" }
        createdAt: { type: string, format: date-time }
...
    ConnectorField:
      type: object
      required: [key, label, secret, optional, events]
      properties:
        key: { type: string }
        label: { type: string }
        help: { type: string }
        secret: { type: boolean }
        optional: { type: boolean }
        events:
          type: boolean
          description: |
            Only needed to receive events (e.g. a webhook signing secret). Filled in after
            connecting, next to the webhook URL, rather than in the first form.
...
    ConnectorSetup:
      type: object
      required: [steps, eventSteps]
      properties:
        steps:
          type: array
          description: How to get the credentials to connect
          items: { $ref: "#/components/schemas/SetupStep" }
        eventSteps:
          type: array
          description: How to start receiving events once connected (with the webhook URL)
          items: { $ref: "#/components/schemas/SetupStep" }
    ConnectorSignalType:
      type: object
      required: [type, description, fields]
      properties:
        type: { type: string }
        description: { type: string }
        fields:
          type: array
          items: { type: string }
    ConnectorType:
      type: object
      required: [type, name, description, credentialFields, configFields, signals, webhooks, setup]
      properties:
        type: { type: string, description: 'Machine name, e.g. "github"' }
        name: { type: string, description: 'Display name, e.g. "GitHub"' }
        description: { type: string }
        credentialFields:
          type: array
          items: { $ref: "#/components/schemas/ConnectorField" }
        configFields:
          type: array
          items: { $ref: "#/components/schemas/ConnectorField" }
        signals:
          type: array
          items: { $ref: "#/components/schemas/ConnectorSignalType" }
        webhooks:
          type: boolean
          description: Receives events at a webhook URL (shown after connecting)
        setup: { $ref: "#/components/schemas/ConnectorSetup" }
    Connector:
      type: object
      required: [id, type, name, config, credentialsSet, agentIds, webhookUrl, webhookSecret, signIn, createdAt]
      properties:
        id: { type: string }
        type: { type: string }
        name: { type: string }
        config:
          type: object
          additionalProperties: { type: string }
        credentialsSet:
          type: array
          description: Credential keys that have a saved value (values are never returned)
          items: { type: string }
        agentIds:
          type: array
          description: Agents allowed to use this account
          items: { type: string }
        webhookUrl:
          type: string
          nullable: true
          description: Where the service should send events (types with webhooks)
        webhookSecret:
          type: string
          nullable: true
          description: |
            A secret openbot generated for the service's webhook settings (e.g. GitHub's), shown so
            the user can copy it there
        signIn:
          type: string
          nullable: true
          enum: [ok, expired]
          description: Connected by signing in (MCP OAuth); expired means it needs Reconnect
        createdAt: { type: string, format: date-time }
    CreateConnectorRequest:
      type: object
      required: [type, name, credentials]
      properties:
        type: { type: string }
        name: { type: string, minLength: 1, maxLength: 64 }
        credentials:
          type: object
          additionalProperties: { type: string }
        config:
          type: object
          additionalProperties: { type: string }
        agentIds:
          type: array
          items: { type: string }
    UpdateConnectorRequest:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 64 }
        credentials:
          type: object
          additionalProperties: { type: string }
        config:
          type: object
          additionalProperties: { type: string }
        agentIds:
          type: array
          items: { type: string }
...
    AgentActivity:
      type: object
      required: [state, label]
      properties:
        state: { type: string, enum: [idle, working] }
        label:
          type: string
          nullable: true
          description: What the agent is doing, e.g. "thinking" or "running npm test"
...
    MessageAuthor:
      type: object
      required: [kind, agentId]
      properties:
        kind: { type: string, enum: [user, agent, system] }
        agentId:
          type: string
          nullable: true
          description: The agent that wrote it; null on an agent message means the agent was deleted
    Message:
      type: object
      required: [id, chatId, author, body, createdAt, reactions, mentions, attachments]
      properties:
        id:
          type: string
          description: ULID; lexicographic order is chronological order
        chatId: { type: string }
        author: { $ref: "#/components/schemas/MessageAuthor" }
        body:
          type: string
          description: Markdown
        createdAt: { type: string, format: date-time }
        failure:
          allOf: [{ $ref: "#/components/schemas/MessageFailure" }]
          nullable: true
          description: Set on system messages reporting that an agent's turn failed
        mentions:
          type: array
          description: Agents @mentioned in the message (group chats)
          items: { type: string }
        attachments:
          type: array
          description: Files and images attached to the message, in order
          items: { $ref: "#/components/schemas/Attachment" }
        reactions:
          type: array
          description: Emoji reactions, in the order each emoji was first used
          items: { $ref: "#/components/schemas/Reaction" }
        prompt:
          allOf: [{ $ref: "#/components/schemas/Prompt" }]
          nullable: true
          description: |
            Set on agent messages that ask the user something with clickable answers. `body`
            holds the question as plain text (for previews).
        event:
          allOf: [{ $ref: "#/components/schemas/MessageEvent" }]
          nullable: true
          description: Set on system messages that mark something that happened, e.g. an agent was created
        clientId:
          type: string
          description: |
            Echo of SendMessageRequest.clientId. Only present on the sendMessage response and the
            matching message.created event.
    MessageFailure:
      type: object
      required: [agentId, reason, retryable]
      properties:
        agentId: { type: string }
        reason:
          type: string
          enum: [no_key, invalid_key, model_blocked, provider_error, too_many_steps]
          description: |
            model_blocked: OpenCode refuses the agent's model (its provider trains on request
            data and the workspace's privacy settings forbid that). Offer to change the model.
        retryable: { type: boolean }
...
    Reaction:
      type: object
      required: [emoji, by]
      properties:
        emoji: { type: string }
        by:
          type: array
          description: Who reacted with this emoji, oldest first
          items: { $ref: "#/components/schemas/MessageAuthor" }
    Prompt:
      type: object
      required: [kind, question, options, allowOther, status, answer]
      properties:
        kind:
          type: string
          enum: [single, multi, text, approval, connect]
          description: |
            single = pick one; multi = pick any number; text = free-text answer; approval = the
            agent wants to do something that needs the user's OK (options: Approve, Decline, and
            sometimes a third, "Always allow", which approves and stops asking for that action);
            connect = the agent proposes connecting an app (see connection; options: Connect,
            Decline; connect with POST /messages/{messageId}/connect)
        question: { type: string }
        options:
          type: array
          items: { $ref: "#/components/schemas/PromptOption" }
        allowOther:
          type: boolean
          description: Offer "type your own" in addition to the options
        status: { type: string, enum: [pending, answered, dismissed] }
        answer:
          allOf: [{ $ref: "#/components/schemas/PromptAnswer" }]
          nullable: true
        connection:
          allOf: [{ $ref: "#/components/schemas/PromptConnection" }]
          nullable: true
          description: The proposed connection (connect prompts only)
        preview:
          allOf: [{ $ref: "#/components/schemas/ActionPreview" }]
          nullable: true
          description: What an approval prompt would do, for the card (approval prompts only)
...
    ConnectPromptRequest:
      type: object
      required: [credentials]
      properties:
        credentials:
          type: object
          additionalProperties: { type: string }
          description: Values for the type's credential fields
        name:
          type: string
          description: Overrides the proposed name
    PromptOption:
      type: object
      required: [label]
      properties:
        label: { type: string }
        opensChatId:
          type: string
          description: Choosing this option also opens this chat (e.g. "Go talk to X")
...
    Attachment:
      type: object
      required: [id, name, mime, size, url, width, height]
      properties:
        id: { type: string }
        name: { type: string, description: The file name }
        mime: { type: string, description: 'e.g. "image/png", "application/pdf"' }
        size: { type: integer, format: int64, description: Bytes }
        url: { type: string, description: Where to fetch it (same origin, needs the session) }
        width: { type: integer, nullable: true, description: Pixels, for images }
        height: { type: integer, nullable: true, description: Pixels, for images }
...
security:
  - session: []